requirements.txt
apache-airflow==2.2.3
scikit-learn==0.24.2
mlflow==1.20.2
pandas==1.3.0
joblib==1.1.0
